+++ src/api/objects/account.ts
@@ -7,7 +7,6 @@ import { AccountNames } from "../../types";
 import { extractDomain, replaceEmojiShortcodesWithImageTags } from "../../helpers/string_helpers";
 import { keyByProperty } from "../../helpers/collection_helpers";
 import { MastoApi } from "../api";
-import { key } from "localforage";
 
 interface AccountObj extends mastodon.v1.Account {
     describe?: () => string;